+++ ding/model/common/utils.py
@@ -23,16 +23,25 @@ def create_model(cfg: EasyDict) -> torch.nn.Module:
     return MODEL_REGISTRY.build(cfg.pop("type"), **cfg)
 
 
-def top_p_logits(logits, topp=0.9, filter_value=0, min_topk=1):
+def top_p_logits(logits: torch.Tensor, topp: float = 0.9, filter_value: float = 0, min_topk: int = 1):
     """
-    Filter a distribution of logits using nucleus (top-p) filtering
-    https://github.com/OpenLMLab/MOSS/blob/e088f438d1a95d424c6dffef0d73134ebe62cb72/models_jittor/generation.py#L146
+    Overview:
+        Filter a distribution of logits using nucleus (top-p) filtering. The output is also logit tensors but some \
+        values are masked.
+    Arguments:
+        - logits (:obj:`torch.Tensor`): The input logits for top-p sampling.
+        - topp (:obj:`float`): The top-p value, such as 0.9.
+        - filter_value (:obj:`float`): The value for masked logits in output, default as 0.
+        - min_topk (:obj:`int`): The min number of sampled logit, default as 1 (which means that at least one sample \
+            will not be masked.)
+    Returns:
+        - cum_logits (:obj:`torch.Tensor`): The output logits after masking.
     """
     cum_logits = logits.clone()
     if topp > 0:
         logits_sorted, inds = torch.sort(logits, dim=-1, descending=True)
         mask = (logits_sorted.cumsum(dim=-1) - logits_sorted) >= topp
-        mask[:, :min_topk] = False
+        mask[..., :min_topk] = False
         # Remove tokens with cumulative top_p above the threshold
         mask = torch.zeros_like(mask).to(torch.bool).scatter_(dim=-1, index=inds, src=mask)
         cum_logits[mask] = filter_value
